Change #2
- Implemented music playing functionality via Minim
- Added all music-related things to its own tab for organizational
reasons (though it is not a class, nor likely will be)
- Added all major and harmonic minor scales as choosable options though
a 2D array
- Attempted to play around with octave switching, though not sure how
to proceed with this (octave switching would require reinitializing the
entire 2D array, but this cannot be done)
- Implemented chords by playing 3 notes at once, enabled as the default
(can be disabled by pressing c, and will apply to the next iteration of
notes)
- Removed all image-related traces from the main project, save for the
obvious Kinect ones
- Made a few minor changes here and there, cleaned up some code
(deleted rotation in the entity class, for instance, changed
friction…). Some changes may have been simply unmentioned from last
commit.
- Setup a millis timer to switch tracks
- Implemented a recording function